added alot of comments and did some cleanup of src
