+++ src/Plugin/Field/ReactParagraphsFields/Ckeditor.php
@@ -10,7 +10,7 @@ use Drupal\react_paragraphs\ReactParagraphsFieldsBase;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Class Text
+ * Ckeditor field plugin.
  *
  * @ReactParagraphsFields(
  *   id = "ckeditor",
@@ -23,10 +23,15 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 class Ckeditor extends ReactParagraphsFieldsBase implements ContainerFactoryPluginInterface {
 
   /**
+   * Current user object.
+   *
    * @var \Drupal\Core\Session\AccountProxyInterface
    */
   protected $currentUser;
 
+  /**
+   * {@inheritDoc}
+   */
   public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
     return new static(
       $configuration,
@@ -36,6 +41,9 @@ class Ckeditor extends ReactParagraphsFieldsBase implements ContainerFactoryPlug
     );
   }
 
+  /**
+   * {@inheritDoc}
+   */
   public function __construct(array $configuration, $plugin_id, $plugin_definition, AccountProxyInterface $current_user) {
     parent::__construct($configuration, $plugin_id, $plugin_definition);
     $this->currentUser = $current_user;
