+++ src/Plugin/Field/ReactParagraphsFields/Datetime.php
@@ -6,7 +6,7 @@ use Drupal\field\FieldConfigInterface;
 use Drupal\react_paragraphs\ReactParagraphsFieldsBase;
 
 /**
- * Class Text
+ * Datetime and daterange field plugin.
  *
  * @ReactParagraphsFields(
  *   id = "datetime",
