+++ src/Plugin/Field/ReactParagraphsFields/EntityReference.php
@@ -9,7 +9,7 @@ use Drupal\react_paragraphs\ReactParagraphsFieldsBase;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Class Text
+ * Entity reference field plugin.
  *
  * @ReactParagraphsFields(
  *   id = "entity_reference",
@@ -22,10 +22,15 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 class EntityReference extends ReactParagraphsFieldsBase implements ContainerFactoryPluginInterface {
 
   /**
+   * Entity type manager service.
+   *
    * @var \Drupal\Core\Entity\EntityTypeManagerInterface
    */
   protected $entityTypeManager;
 
+  /**
+   * {@inheritDoc}
+   */
   public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
     return new static(
       $configuration,
@@ -35,6 +40,9 @@ class EntityReference extends ReactParagraphsFieldsBase implements ContainerFact
     );
   }
 
+  /**
+   * {@inheritDoc}
+   */
   public function __construct(array $configuration, $plugin_id, $plugin_definition, EntityTypeManagerInterface $entity_type_manager) {
     parent::__construct($configuration, $plugin_id, $plugin_definition);
     $this->entityTypeManager = $entity_type_manager;
@@ -47,19 +55,29 @@ class EntityReference extends ReactParagraphsFieldsBase implements ContainerFact
     $info = parent::getFieldInfo($field_element, $field_config);
 
     if (isset($field_element['widget']['media_library_selection'])) {
-      $this->addMediaLibraryInfo($info, $field_element, $field_config);
+      $this->addMediaLibraryInfo($info, $field_config);
     }
     return $info;
   }
 
-  protected function addMediaLibraryInfo(&$info, $field_element, FieldConfigInterface $field_config) {
+  /**
+   * When the widget is a media library, add data to the field info.
+   *
+   * @param array $info
+   *   Field info data array.
+   * @param \Drupal\field\FieldConfigInterface $field_config
+   *   Media library field.
+   *
+   * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
+   * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
+   */
+  protected function addMediaLibraryInfo(array &$info, FieldConfigInterface $field_config) {
     $info['widget_type'] = 'media_library';
     $target_bundles = $this->entityTypeManager->getStorage('media_type')
       ->loadMultiple($field_config->getSetting('handler_settings')['target_bundles']);
     $info['target_bundles'] = array_map(function ($bundle) {
       return $bundle->label();
     }, $target_bundles);
-
   }
 
 }
