+++ src/Plugin/Field/ReactParagraphsFields/Number.php
@@ -6,7 +6,7 @@ use Drupal\field\FieldConfigInterface;
 use Drupal\react_paragraphs\ReactParagraphsFieldsBase;
 
 /**
- * Class Text
+ * Number field plugin.
  *
  * @ReactParagraphsFields(
  *   id = "number",
