+++ src/Plugin/Field/ReactParagraphsFields/View.php
@@ -6,7 +6,7 @@ use Drupal\field\FieldConfigInterface;
 use Drupal\react_paragraphs\ReactParagraphsFieldsBase;
 
 /**
- * Class Text
+ * Viewfield Plugin.
  *
  * @ReactParagraphsFields(
  *   id = "view",
